Move callback classes to callbacks module
